+++ src/entities/post/model/PostContext.tsx
@@ -1,7 +1,7 @@
 import { createContext, PropsWithChildren, useContext, useMemo } from "react"
-import { usePosts } from "../../../features/post-view/model/usePosts"
 import { Post } from "../model/types"
 import { User } from "../../user/model/types"
+import { usePosts } from "./usePosts"
 
 interface PostContextType {
   posts: Post[]
