+++ src/pages/PostsManagerPage.tsx
@@ -6,9 +6,9 @@ import { PostDetailModal } from "../features/post-detail/ui/PostDetailModal"
 import { UserModal } from "../features/user/ui/UserModal"
 import { PostAddModal } from "../features/post-add/ui/PostAddModal"
 import { usePostsContext } from "../entities/post/model/PostContext"
-import { PostTable } from "../features/post-view/ui/PostTable"
-import { Pagination } from "../features/post-view/ui/Pagination"
-import { SelectTag } from "../features/tag-select/ui/TagSelect"
+import { PostTable } from "../widgets/post/ui/PostTable"
+import { Pagination } from "../features/post-pagination/ui/Pagination"
+import { SelectTag } from "../features/post-filter/ui/TagSelect"
 import { SelectSortStandard } from "../features/post-sort/ui/SelectSortStandard"
 import { SelectSortOrder } from "../features/post-sort/ui/SelectSortOrder"
 import { PostEditModal } from "../features/post-edit/ui/PostEditModal"
