+++ src/widgets/post/ui/PostTable.tsx
@@ -1,0 +1,20 @@
+import { Table, TableBody } from "../../../shared/ui"
+import { PostTableHead } from "./PostTableHead"
+import { usePostsContext } from "../../../entities/post/model/PostContext"
+import { PostItem } from "../../../features/post-item/ui/PostItem"
+
+export const PostTable = () => {
+  const { posts } = usePostsContext()
+
+  return (
+    <Table>
+      <PostTableHead />
+
+      <TableBody>
+        {posts.map((post) => (
+          <PostItem key={post.id} post={post} />
+        ))}
+      </TableBody>
+    </Table>
+  )
+}
